integrate get section data into one func

+++ src/runtime_src/driver/xclng/drm/xocl/userpf/common.h
@@ -158,7 +158,7 @@ int xocl_init_sysfs(struct device *dev);
 void xocl_fini_sysfs(struct device *dev);
 
 /* helper functions */
-int xocl_hot_reset(struct xocl_dev *xdev, bool force);
+int64_t xocl_hot_reset(struct xocl_dev *xdev, bool force);
 void xocl_p2p_mem_release(struct xocl_dev *xdev, bool recov_bar_sz);
 int xocl_p2p_mem_reserve(struct xocl_dev * xdev);
 int xocl_get_p2p_bar(struct xocl_dev *xdev, u64 *bar_size);
